Removed (most?) remaining references to IncidentReport
