+++ manage.py
@@ -16,7 +16,7 @@ from app.models import (
     User,
     Role,
     Permission,
-    IncidentReport,
+    Incident,
     EditableHTML
 )
 
@@ -78,7 +78,7 @@ def add_fake_data(number_users, number_reports):
     Adds fake data to the database.
     """
     User.generate_fake(count=number_users)
-    IncidentReport.generate_fake(count=number_reports)
+    Incident.generate_fake(count=number_reports)
 
 
 @manager.command
